Move signal_to_raw_data() to KomaMRICore
