+++ KomaMRIIO/test/runtests.jl
@@ -24,29 +24,6 @@ using TestItems, TestItemRunner
         seq = @suppress read_seq(path*"/test_files/radial_JEMRIS.seq") #Pulseq v1.2.1
         @test seq.DEF["FileName"] == "radial_JEMRIS.seq"
         @test seq.DEF["PulseqVersion"] ≈ 1002001
-
-        # Test ISMRMRD
-        fraw = ISMRMRDFile(path*"/test_files/Koma_signal.mrd")
-        raw = RawAcquisitionData(fraw)
-        @test raw.params["protocolName"] == "epi"
-        @test raw.params["institutionName"] == "Pontificia Universidad Catolica de Chile"
-        @test raw.params["encodedSize"] ≈ [101, 101, 1]
-        @test raw.params["reconSize"] ≈ [102, 102, 1]
-        @test raw.params["patientName"] == "brain2D_axial"
-        @test raw.params["trajectory"] == "other"
-        @test raw.params["systemVendor"] == "KomaMRI.jl"
-
-        # Test signal_to_raw_data
-        signal1 = Vector()
-        for i=1:length(raw.profiles)
-            signal1 = [signal1; raw.profiles[i].data]
-        end
-        rawmrd = signal_to_raw_data(signal1, seq)
-        @test rawmrd.params["institutionName"] == raw.params["institutionName"]
-        io = IOBuffer()
-        show(io, "text/plain", rawmrd)
-        @test occursin("RawAcquisitionData[", String(take!(io)))
-
     end
     # Test JEMRIS
     @testset "JEMRIS" begin
